LayoutTests/imported/w3c/web-platform-tests/2dcontext/fill-and-stroke-styles/2d.fillStyle.parse.current.removed-expected.txt

2d.fillStyle.parse.current.removed
currentColor is solid black when the canvas element is not in a document
Actual output:
Expected output:


PASS currentColor is solid black when the canvas element is not in a document 
